LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/exclude-fixed-position-expected.txt

fixed
content

FAIL Fixed-position header. assert_equals: expected 150 but got 100
